requirements.txt
asgiref==3.5.2
black==22.1.0
blue==0.9.0
click==8.1.3
Django==4.0.6
flake8==4.0.1
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0

djangorestframework~=3.13.1
PyMySQL~=1.0.2
PyYAML~=6.0